fix(build): resolve all compilation errors across preview, social hub and download dialog v3.0.7

+++ app/src/main/java/com/junkfood/seal/ui/component/MediaPreviewDialog.kt
@@ -232,7 +232,7 @@ fun MediaPreviewDialog(
                             }
                             Spacer(modifier = Modifier.height(8.dp))
                             Text(
-                                text = file.extension.uppercase(Locale.US) + " • " + FileUtil.formatFileSize(file.length().toDouble()),
+                                text = file.extension.uppercase(Locale.US) + " • " + FileUtil.formatFileSize(file.length()),
                                 style = MaterialTheme.typography.labelSmall.copy(fontWeight = FontWeight.Bold),
                                 color = MaterialTheme.colorScheme.primary
                             )
